requirements.txt
anyio==3.5.0
base58==2.1.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
ECPy==1.2.5
Flask==2.0.3
h11==0.12.0
httpcore==0.13.7
httpx==0.18.2
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
python-dotenv==0.19.2
requests==2.27.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
typing-extensions==3.10.0.2
urllib3==1.26.9
websocket-client==1.2.3
websockets==10.2
Werkzeug==2.0.3
xrpl-py==1.4.0
xumm-sdk-py==1.0.1
